% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oap_classes.R
\name{plan_node}
\alias{plan_node}
\title{Fetch a node by id}
\usage{
plan_node(plan, id)
}
\arguments{
\item{plan}{An \link{oap_plan}.}

\item{id}{The id of the node.}
}
\value{
The \link{oap_node} with that id. An unknown id is an error.
}
\description{
Fetch a node by id
}
\examples{
p <- oap_plan(id = "one", nodes = list(node("a")))
plan_node(p, "a")@id

}
\author{
Max Moldovan
}
