% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oap_classes.R
\name{plan_order}
\alias{plan_order}
\title{Topological order of a plan}
\usage{
plan_order(plan)
}
\arguments{
\item{plan}{An \link{oap_plan}.}
}
\value{
A character vector of node ids in execution order.
}
\description{
Orders the nodes so that every node follows the nodes it consumes. A plan
with a cycle is an error.
}
\examples{
p <- oap_plan(id = "chain", nodes = list(
  node("b", consumes = list(edge("a"))),
  node("a")
))
plan_order(p)

}
\author{
Max Moldovan
}
