% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductor.R
\name{useConductor}
\alias{useConductor}
\alias{use_conductor}
\title{Dependencies}
\usage{
useConductor()

use_conductor()
}
\description{
Include dependencies, place anywhere in the shiny UI.
}
\examples{
library(shiny)
library(conductor)

ui <- fluidPage(
 useConductor()
 # also works:
 # use_conductor()
)

server <- function(input, output){}

if(interactive()) shinyApp(ui, server)
}
